Revert "Issue #216 resolution"
