+++ frontend/src/views/utente/Preference.js
@@ -7,7 +7,7 @@ import {
   MDBContainer,
   MDBRow, MDBTable, MDBTableBody, MDBTableHead,
 } from "mdb-react-ui-kit";
-import PreferencesDatePick from "../../components/common/PreferencesDatePick";
+import DatePick from "../../components/common/DatePick";
 import IconButton from "@mui/material/IconButton";
 import {DesiderateAPI} from "../../API/DesiderataAPI";
 import {toast, ToastContainer} from "react-toastify";
@@ -27,25 +27,17 @@ export default class Preference extends React.Component {
     super(props);
     this.state = {
       desiderate:[],
-      toDeletePreferences: [],
       orderBy: "data",
       comparator: defaultComparator
     }
     this.setOrderBy = this.setOrderBy.bind(this);
-    this.updatePreferences = this.updatePreferences.bind(this) ;
+
   }
 
   setOrderBy(userProp){
     this.setState({
       orderBy: userProp,
       comparator: defaultComparator
-    })
-  }
-
-  updatePreferences(prefs, toDelPrefs) {
-    this.setState({
-      desiderate : prefs,
-      toDeletePreferences : toDelPrefs,
     })
   }
 
@@ -109,10 +101,9 @@ export default class Preference extends React.Component {
       <MDBContainer className="py-5" style={{height: '85vh',}}>
         <MDBCard alignment='center'>
           <MDBCardBody>
-            <MDBCardTitle>Inserisci le tue desiderata</MDBCardTitle>
+            <MDBCardTitle>Inserisci le tue desiderata   <DatePick onSelectdate={() => this.componentDidMount()} desiderate={this.state.desiderate}/></MDBCardTitle>
             <MDBRow>
             <MDBCol>
-              <PreferencesDatePick onSelectdate={() => this.componentDidMount()} desiderate={this.state.desiderate} toDelPrefs = {this.state.toDeletePreferences} setDesiderate={this.updatePreferences}/>
             </MDBCol>
         </MDBRow>
             <MDBRow>
